Website improvements (#38)
* Comment the parts with the official mobile apps

Regarding this issue: https://github.com/todotxt/todotxt.github.io/issues/35 and this blog post: https://blog.todotxt.org/post/166157430787/todotxt-for-android-and-ios-stopped-syncing-to the official mobile-apps are not available anymore (due to an issue with the Dropbox-API). Let's remove them until this issue is fixed.

* Update Copyright to 2020

* Remove Broken Links

Links didn't work anymore, so i removed them.